Move country related stuff to country sub-feature
