+++ src/features/map/country/init.ts
@@ -9,21 +9,33 @@ import { combine, forward, guard, sample } from 'effector';
 
 import { mapCountry } from '~/core/routes';
 import {
+  fetchCountriesData,
+  fetchCountriesGeometryData,
+} from '~/features/map/api';
+import { combineCountriesDataToGeoJson } from '~/features/map/map-data-helpers';
+import { $map, $stylePaintData, changeMap } from '~/features/map/model';
+import { getInverted, setPayload } from '~/lib/effector-kit';
+
+import {
   $countriesData,
   $countriesGeoJson,
   $countriesGeometryData,
-  $map,
-  $stylePaintData,
-  changeCountryId,
-} from '~/features/map/model';
-import { getInverted } from '~/lib/effector-kit';
-
-import {
+  $selectedCountryId,
   addCountriesFx,
+  changeCountryId,
+  fetchCountriesDataFx,
+  fetchCountriesGeometryDataFx,
   leaveCountryRouteFx,
   updateCountryFx,
   updateSchoolsFx,
 } from './model';
+
+fetchCountriesDataFx.use(fetchCountriesData);
+fetchCountriesGeometryDataFx.use(fetchCountriesGeometryData);
+
+$countriesData.on(fetchCountriesDataFx.doneData, setPayload);
+$countriesGeometryData.on(fetchCountriesGeometryDataFx.doneData, setPayload);
+$selectedCountryId.on(changeCountryId, setPayload);
 
 const $changeCountryData = combine({
   map: $map,
@@ -40,6 +52,20 @@ forward({
   to: [updateCountryFx, updateSchoolsFx],
 });
 
+// Routing
+sample({
+  source: guard(mapCountry.params, { filter: Boolean }),
+  fn: (params) => Number(params?.id),
+  target: changeCountryId,
+});
+
+sample({
+  source: mapCountry.params,
+  clock: changeMap,
+  fn: (params) => (params?.id ? Number(params.id) : 0),
+  target: changeCountryId,
+});
+
 // Leave country route
 sample({
   source: $changeCountryData,
@@ -49,15 +75,20 @@ sample({
   target: leaveCountryRouteFx,
 });
 
-// Add countries
-const onCountriesGeoJson = sample({
-  source: $countriesGeoJson,
-  clock: guard({
-    source: combine([$countriesData, $countriesGeometryData]),
-    filter: ([countriesData, countriesGeometryData]) =>
-      Boolean(countriesData && countriesGeometryData),
-  }),
+const allCountriesDataLoaded = guard({
+  source: combine([$countriesData, $countriesGeometryData]),
+  filter: ([countriesData, countriesGeometryData]) =>
+    Boolean(countriesData && countriesGeometryData),
 });
+
+$countriesGeoJson.on(
+  allCountriesDataLoaded,
+  (_, [countriesData, countriesGeometryData]) =>
+    combineCountriesDataToGeoJson(countriesData, countriesGeometryData)
+);
+
+// Add countries
+const onCountriesGeoJson = sample($countriesGeoJson, allCountriesDataLoaded);
 
 sample({
   source: $changeCountryData,
