+++ src/features/map/country/types.ts
@@ -1,4 +1,4 @@
-import { FeatureCollection } from 'geojson';
+import { FeatureCollection, Geometry } from 'geojson';
 
 import { Map, StylePaintData } from '~/features/map/types';
 
@@ -22,4 +22,25 @@ export type AddCountries = {
   map: Map | null;
   paintData: StylePaintData;
   countriesGeoJson: FeatureCollection;
+};
+
+export type CountryData = {
+  id: number;
+  name: string;
+  code: string;
+  flag: string;
+  map_preview: null | string;
+  description: null | string;
+  data_source: null | string;
+  integration_status: IntegrationStatus;
+  date_of_join: null | string;
+  schools_with_data_percentage: null | string;
+  geometry?: Geometry;
+};
+
+export type IntegrationStatus = 0 | 1 | 2 | 3;
+
+export type CountryGeometryData = {
+  id: number;
+  geometry_simplified: Geometry;
 };
