+++ src/features/map/index.ts
@@ -1,4 +1,3 @@
 import './init';
 
 export { Map } from './map';
-export { fetchCountriesDataFx } from './model';
